cast int to long long to silence warning

+++ src/type_hindexed_x.c
@@ -41,7 +41,7 @@ int MPIX_Type_create_hvector_x(int count,
 
     /* The count has to fit into MPI_Aint for BigMPI to work. */
     if ((MPI_Count)count>(MPI_Count)bigmpi_count_max) {
-        printf("count (%lld) exceeds bigmpi_count_max (%lld)\n", count, bigmpi_count_max);
+        printf("count (%lld) exceeds bigmpi_count_max (%lld)\n", (long long)count, bigmpi_count_max);
         fflush(stdout);
     }
